Add files via upload
This paper for an Econometric course during my fist year of Master Degree studies how changes in the ECB policy rate relate to exchange rates of the USD, JPY and GBP. Framed by interest‑rate parity, it contrasts the USD as a global reserve currency, the JPY as a safe haven with ultra‑accommodative policy, and the GBP as a non‑euro European currency. The analysis is preliminary since only the ECB rate is used (foreign rates are not directly included), but relies on very recent data and compares results across currency zones.